fixdesk/requirements.txt
amqp==5.3.1
asgiref==3.11.0
asttokens==3.0.1
billiard==4.2.4
celery==5.6.2
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
comm==0.2.3
cryptography==46.0.5
debugpy==1.8.19
decorator==5.2.1
defusedxml==0.7.1
dj-database-url==3.1.0
Django==6.0.1
django-cors-headers==4.9.0
django-filter==25.2
django-redis==6.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
executing==2.2.1
gunicorn==23.0.0
h11==0.16.0
idna==3.11
ipykernel==7.1.0
ipython==9.9.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jupyter_client==8.8.0
jupyter_core==5.9.1
kombu==5.6.2
matplotlib-inline==0.2.1
nest-asyncio==1.6.0
numpy==2.4.2
oauthlib==3.3.1
packaging==25.0
pandas==3.0.0
parso==0.8.5
platformdirs==4.5.1
prompt_toolkit==3.0.52
psutil==7.2.2
psycopg2==2.9.11
pure_eval==0.2.3
pycparser==3.0
Pygments==2.19.2
PyJWT==2.11.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python3-openid==3.2.0
pyzmq==27.1.0
rabbitmq==0.2.0
redis==7.1.0
requests==2.32.5
requests-oauthlib==2.0.0
six==1.17.0
social-auth-app-django==5.7.0
social-auth-core==4.8.5
sqlparse==0.5.5
stack-data==0.6.3
tornado==6.5.4
traitlets==5.14.3
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
uvicorn==0.40.0
vine==5.1.0
waitress==3.0.2
wcwidth==0.5.0
whitenoise==6.11.0
